requirements.txt
matplotlib==3.8.0
numpy==1.25.2
opacus==1.4.0
Pillow==9.4.0
tensorboardX==2.6.2.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
